Fix header-dropdown-nav
