+++ src/styles/index.pcss
@@ -7,3 +7,11 @@
   padding-left: 3px;
   vertical-align: 30%;
 }
+
+.header-brand > :not(.nav-item) {
+  display: contents;
+}
+
+.display-contents {
+  display: contents;
+}
