+++ src/templates/index.posthtml
@@ -18,22 +18,13 @@
 
     <include src="partials/firebase-sdk.html"></include>
 
-    <div class="header header-fixed unselectable header-animated">
+    <div id="header" class="header header-fixed unselectable header-animated">
       <div class="header-brand">
         <div class="nav-item no-hover">
           <a><h5 class="title">Talker</h5></a>
         </div>
-        <div class="nav-item nav-btn" id="header-btn">
-          <span>First Child</span>
-          <span>Second Child</span>
-          <span>Third Child</span>
-        </div>
       </div>
-      <div class="header-nav" id="header-menu">
-        <div class="nav-right">
-          <div class="nav-item has-sub" id="dropdown-click">
-          </div>
-        </div>
+      <div id="header-menu" class="display-contents">
       </div>
     </div>
